// typscript/InterfaceStrava.d.ts
export interface StravaToken {
  _id: string;
  athlete_id: number;
  access_token: string;
  athlete: object;
  expires_at: number;
  expires_in: number;
  refresh_token: string;
  token_type: string;
}

export interface StravaActivity {
  _id: string;
  id: number;
  achievement_count: number;
  athlete: object
  athlete_count: number;
  available_zones: Array<any>;
  average_cadence: number;
  average_heartrate: number;
  average_speed: number;
  average_temp: number;
  best_efforts: Array<any>;
  calories: number;
  comment_count: number;
  commute: boolean;
  description: string;
  device_name: string;
  display_hide_heartrate_option: boolean;
  distance: number;
  elapsed_time: number;
  elev_high: number;
  elev_low: number;
  embed_token: string;
  end_latlng: Array<any>;
  external_id: string;
  flagged: boolean;
  from_accepted_tag: boolean;
  gear_id: null | string;
  has_heartrate: boolean;
  has_kudoed: boolean;
  heartrate_opt_out: boolean;
  hide_from_home: boolean;
  kudos_count: number;
  laps: Array<any>;
  location_city: null | string;
  location_country: null | string;
  location_state: null | string;
  manual: boolean;
  map?: string;
  max_heartrate: number;
  max_speed: number;
  moving_time: number;
  name: string;
  perceived_exertion: any;
  photo_count: number;
  photos: object;
  pr_count: number;
  prefer_perceived_exertion: any;
  private: boolean;
  resource_state: number;
  segment_efforts: Array<any>;
  similar_activities: object;
  splits_metric: Array<any>;
  splits_standard: Array<any>;
  start_date: string;
  start_date_local: string;
  start_latitude: number;
  start_latlng: Array<any>;
  start_longitude: number;
  stats_visibility: Array<any>;
  timezone: string;
  total_elevation_gain: number;
  total_photo_count: number;
  trainer: boolean;
  type: string;
  upload_id: number;
  upload_id_str: string;
  utc_offset: number;
  visibility: string;
  workout_type: null | string;
}